Menu Fixed

+++ provider-service/provider-service/src/main/java/com/homebite/provider_service/Repositories/ProviderRepo.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface ProviderRepo extends JpaRepository<Provider,Integer> {
+public interface ProviderRepo extends JpaRepository<Provider,Long> {
 
     Optional<Provider> findByEmail(String email);
 }
